refactor(sdk): rename packages/syner to @syner/sdk (#486)
refactor(sdk): rename packages/syner to @syner/sdk (#469)

- Directory: packages/syner/ → packages/sdk/
- Package name: syner → @syner/sdk
- Updates 19 import sites across apps/bot, apps/dev, packages/vercel
- Updates 3 consumer package.json dependencies
