+++ apps/bot/package.json
@@ -23,7 +23,7 @@
     "next": "^15.0.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "syner": "workspace:*",
+    "@syner/sdk": "workspace:*",
     "zod": "^3.23.0"
   },
   "devDependencies": {
