+++ apps/dev/app/api/skills/[slug]/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { getSkillBySlug, getSkillsList } from "syner/skills";
+import { getSkillBySlug, getSkillsList } from "@syner/sdk/skills";
 import path from "path";
 
 // ISR: revalidate every hour
