+++ apps/dev/components/skill-modal.tsx
@@ -10,7 +10,7 @@ import {
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
 import { Copy, Check } from "lucide-react";
-import type { SkillContent } from "syner/skills";
+import type { SkillContent } from "@syner/sdk/skills";
 
 interface SkillModalProps {
   skill: SkillContent | null;
