+++ apps/dev/package.json
@@ -23,7 +23,7 @@
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "swr": "^2.4.1",
-    "syner": "workspace:*",
+    "@syner/sdk": "workspace:*",
     "tailwind-merge": "^3.5.0"
   },
   "devDependencies": {
