--- packages/syner/src/agents/index.ts
@@ -1,3 +1,0 @@
-export { getAgentsRegistry, getAgentsList, getAgentByName, getAgentsByChannel, invalidateAgentsCache } from './loader'
-export type { AgentCard } from './loader'
-export { getModel, MODEL_IDS } from './models'
