--- packages/syner/src/context/index.ts
@@ -1,5 +1,0 @@
-export { type VaultStore } from './vault-store'
-export { FileSystemVaultStore } from './filesystem-vault-store'
-export { BlobVaultStore } from './blob-vault-store'
-export { type ContextScope, type ContextRequest, type Brief } from './types'
-export { resolveContext } from './resolve'
