--- packages/syner/src/logger.ts
@@ -1,25 +1,0 @@
-type LogLevel = 'debug' | 'info' | 'warn' | 'error'
-
-interface LogEntry {
-  timestamp: string
-  level: LogLevel
-  message: string
-  [key: string]: unknown
-}
-
-export function log(level: LogLevel, message: string, meta?: Record<string, unknown>): void {
-  const entry: LogEntry = {
-    timestamp: new Date().toISOString(),
-    level,
-    message,
-    ...meta,
-  }
-  console.log(JSON.stringify(entry))
-}
-
-export const logger = {
-  debug: (message: string, meta?: Record<string, unknown>) => log('debug', message, meta),
-  info: (message: string, meta?: Record<string, unknown>) => log('info', message, meta),
-  warn: (message: string, meta?: Record<string, unknown>) => log('warn', message, meta),
-  error: (message: string, meta?: Record<string, unknown>) => log('error', message, meta),
-}
