--- packages/syner/src/skills/client.ts
@@ -1,3 +1,0 @@
-// Client-safe exports (no node dependencies)
-export type { Skill, SkillContent } from './types'
-export { groupByCategory } from './types'
