--- packages/syner/src/skills/index.ts
@@ -1,4 +1,0 @@
-export { getSkillsRegistry, getSkillsList, getSkillBySlug, getCategories, getPublicSkills, getInstanceSkills, getPrivateSkills, invalidateSkillsCache } from './loader'
-export type { Skill, SkillContent, SkillVisibility } from './types'
-export { groupByCategory } from './types'
-export { resolveSkill, type ResolvedSkill } from './resolver'
