+++ packages/vercel/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@syner/osprotocol": "workspace:*",
-    "syner": "workspace:*",
+    "@syner/sdk": "workspace:*",
     "@vercel/sandbox": "^1.8.0",
     "ai": "^6.0.0",
     "glob": "^10.0.0",
