+++ packages/vercel/src/tools/vault-list.ts
@@ -1,6 +1,6 @@
 import { tool } from 'ai'
 import { z } from 'zod'
-import type { VaultStore } from 'syner/context'
+import type { VaultStore } from '@syner/sdk/context'
 
 const inputSchema = z.object({
   pattern: z.string().describe('Glob pattern to match files (e.g. "bot/**/*.md", "**/*.md")'),
